+++ UI/CHANGES.md
@@ -1,5 +1,7 @@
 #### Version: 1.0.0-SNAPSHOT (LibGDX 1.7.1)
+- **Added**: constructor TabbedPane(TabbedPaneStyle style, Sizes sizes)
 - **Added**: VisTextField#toString() now returns field text
+- **Fixed**: TabbedPane: Tab close button too small when using SkinScale.X2
 - **Skin**: 
     - **Added**: Drawables: `grey`, `vis-blue`, `vis-red`